% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opusreader.R
\docType{package}
\name{opusreader}
\alias{opusreader}
\title{Reads OPUS Binary Files in R.}
\description{
\code{opuseader} implements a driver for FTIR spectroscopy data collected in the OPUS binary format.
}
\author{
Philipp Baumann and Pierre Roudier
}
