% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_spectra.0}
\alias{test_spectra.0}
\title{Sample OPUS file containing an MIR soil spectrum}
\format{
A binary OPUS file
}
\description{
Sample OPUS binary file. See Details section for more information.
}
\details{
The spectrum contained in this OPUS file was collected
on a soil sample from New Zealand using a Bruker FTIR spectrometer.
}
\examples{
# Access the location of the ASD file using the following command
fn <- opus_file()
fn
# This function is actually just a shorthand for
fn <- system.file("extdata", "test_spectra.0", package = "opusreader")
fn
}
