% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opus_read_file.R
\name{opus_read}
\alias{opus_read}
\title{Read Bruker OPUS Spectrum Binary Files}
\usage{
opus_read(
  file,
  type = "spec",
  simplify = FALSE,
  wns_digits = 1L,
  progress = TRUE,
  atm_comp_minus4offset = FALSE
)
}
\arguments{
\item{file}{Character vector with path to file(s).}

\item{simplify}{Logical (defaults to \code{FALSE}): if set to \code{TRUE}, returns a flattened list.
The first element of that list (\code{wavenumbers}) is the wavenumbers of the first file read.
The second element (\code{spectra}) is a matrix of the corresponding spectra. Especially useful when
passing more than one file to the \code{file} option, for example to read a suite of spectral file
directly into a matrix.}

\item{wns_digits}{Integer that specifies the number of decimal places used to round
the wavenumbers (values of x-variables) if \code{simplify = TRUE}.}

\item{progress}{Logical (defaults to \code{TRUE}) whether a message is printed when an OPUS binary file
is parsed into an R list entry.}

\item{atm_comp_minus4offset}{Logical whether spectra after atmospheric compensation are read with
an offset of -4 bytes from Bruker OPUS files. Default is \code{FALSE}.}
}
\value{
The output of \code{opus_read()} depends on the value of the \code{simplify} option used in the function call.
\itemize{
\item If \code{simplify = FALSE} (default), \code{opus_read()} returns a list of 10 elements:
\itemize{
\item \code{metadata}: a \code{data.frame} containing metadata from the OPUS file
\item \code{spec} If \code{"spec"} was requested in the \code{type} option, a matrix of the spectrum of the sample (otherwise set to \code{NULL}).
\item \code{spec_no_bc} If \code{"spec_no_bc"} was requested in the \code{type} option, a matrix of the spectrum of the sample without background correction (otherwise set to \code{NULL}).
\item \code{sc_sample} If \code{"sc_sample"} was requested in the \code{type} option, a matrix of the single channel spectrum of the sample (otherwise set to \code{NULL}).
\item \code{sc_ref} If \code{"sc_ref"} was requested in the \code{type} option, a matrix of the single channel spectrum of the reference (otherwise set to \code{NULL}).
\item \code{ig_sample} If \code{"ig_sample"} was requested in the \code{type} option, a matrix of the interferogram of the sample (otherwise set to \code{NULL}).
\item \code{ig_ref}  If \code{"ig_ref"} was requested in the \code{type} option, a matrix of the interferogram of the reference (otherwise set to \code{NULL}).
\item \code{wavenumbers} If \code{"spec"} or \code{"spec_no_bc"} was requested in the \code{type} option, a numeric vector of the wavenumbers of the spectrum of the sample (otherwise set to \code{NULL}).
\item \code{wavenumbers_sc_sample} If \code{"sc_sample"} was requested in the \code{type} option, a numeric vector of the wavenumbers of the single channel spectrum of the sample (otherwise set to \code{NULL}).
\item \code{wavenumbers_sc_ref} If \code{"sc_ref"} was requested in the \code{type} option, a numeric vector of the wavenumbers of the single channel spectrum of the reference (otherwise set to \code{NULL}).
}
\item If \code{simplify = TRUE}, a list of two elements is returned:
\itemize{
\item \code{wavenumbers}: Numeric vector with wavenumbers of the requested spectra.
\item \code{spectra}: Matrix with spectra of requested type (see argument \code{type}).
}
}
}
\description{
Read file(s) acquired with a Bruker Vertex FTIR Instrument.
}
\details{
The type of spectra returned by the function when using \code{type = "spec"} depends on the setting of the Bruker instrument: typically, it can be either absorbance or reflectance.

The type of spectra to extract from the file can also use Bruker's OPUS software naming conventions, as follows:
\itemize{
\item \code{ScSm} corresponds to \code{sc_sample}
\item \code{ScRf} corresponds to \code{sc_ref}
\item \code{IgSm} corresponds to \code{ig_sample}
\item \code{IgRf} corresponds to \code{ig_ref}
}
}
\author{
Philipp Baumann
}
