% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opus_read_file.R
\name{opus_read}
\alias{opus_read}
\title{Read a Bruker OPUS spectrum binary file}
\usage{
opus_read(file, extract = "spc", simple = FALSE, progress = TRUE,
atm_comp_minus4offset = FALSE)
}
\arguments{
\item{file}{Character vector with path to file}

\item{extract}{Character vector of spectra types to extract from OPUS binary
file. Default is \code{"spc"}, which will extract the final spectra, e.g.
expressed in absorbance (named \code{AB} in Bruker OPUS programs). Possible
additional values for the character vector supplied to extract are
\code{"ScSm"} (single channel spectrum of the sample measurement), \
code{"ScRf"} (single channel spectrum of the reference measurement),
\code{"IgSm"} (interferogram of the sample measurement) and \code{"IgRf"}
(interferogram of the reference measurement).}

\item{simple}{Logical (defaults \code{FALSE}): if set to \code{TRUE}, returns a much smaller list. The first object of that list (\code{wavenumbers}) is the wavenumbers of the first file read. The second object (\code{spectra}) is a matrix of the corresponding spectra. Especially useful when passing more than one file to the \code{file} option, for example to read a suite of spectral file directly into a matrix.}

\item{progress}{Logical (defaults \code{TRUE}) whether a message is
printed when an OPUS binary file is parsed into an R list entry.}

\item{atm_comp_minus4offset}{Logical whether spectra after atmospheric
compensation are read with an offset of \code{-4} bites from Bruker OPUS
files. Default is \code{FALSE}.}
}
\description{
Read single binary file acquired with an
Bruker Vertex FTIR Instrument
}
\author{
Philipp Baumann
}
